feature : 게시글 작성, 수정, 삭제 인증

+++ src/main/java/F12/newsfeedproject/api/board/controller/ApiBoardController.java
@@ -1,13 +1,17 @@
 package F12.newsfeedproject.api.board.controller;
 
 import F12.newsfeedproject.api.board.dto.request.BoardRequestDto;
-import F12.newsfeedproject.api.board.dto.response.BoardResponseDto;
 import F12.newsfeedproject.api.board.dto.request.BoardUpdateRequestDto;
+import F12.newsfeedproject.api.board.dto.response.BoardResponseDto;
 import F12.newsfeedproject.api.board.service.ApiBoardService;
+import F12.newsfeedproject.domain.user.entity.User;
+import F12.newsfeedproject.global.exception.member.UnAuthorizedModifyException;
+import F12.newsfeedproject.global.security.UserDetailsImpl;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -26,9 +30,18 @@ public class ApiBoardController {
 
   // 게시글 작성
   @PostMapping
-  public ResponseEntity<BoardResponseDto> saveBoard(@RequestBody BoardRequestDto requestDto) {
-    BoardResponseDto responseDto = boardService.saveBoard(requestDto);
-    return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
+  public ResponseEntity<BoardResponseDto> saveBoard(
+      @RequestBody BoardRequestDto requestDto,
+      @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+    if (userDetails != null) {
+      Long userId = userDetails.getUser().getUserId();
+
+      BoardResponseDto responseDto = boardService.saveBoard(requestDto, userId);
+      return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
+    } else {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    }
   }
 
   // 게시글 단건 조회
@@ -48,16 +61,35 @@ public class ApiBoardController {
   // 게시글 수정
   @PatchMapping("/{boardId}")
   public ResponseEntity<BoardResponseDto> updateBoard(
-      @PathVariable Long boardId, @RequestBody BoardUpdateRequestDto requestDto) {
+      @PathVariable Long boardId, @RequestBody BoardUpdateRequestDto requestDto,
+      @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+    User user = userDetails.getUser();
+    if (!haveModifyAuthorization(user, boardId)) {
+      throw new UnAuthorizedModifyException();
+    }
+
     BoardResponseDto responseDto = boardService.updateBoard(boardId, requestDto);
     return ResponseEntity.ok(responseDto);
   }
 
   // 게시글 삭제
   @DeleteMapping("/{boardId}")
-  public ResponseEntity<Void> deleteBoard(@PathVariable Long boardId) {
+  public ResponseEntity<Void> deleteBoard(@PathVariable Long boardId,
+        @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+    User user = userDetails.getUser();
+    if (!haveModifyAuthorization(user, boardId)) {
+      throw new UnAuthorizedModifyException();
+    }
+
     boardService.deleteBoard(boardId);
     return ResponseEntity.noContent().build();
+  }
+
+  public boolean haveModifyAuthorization(User loginUser, Long boardId) {
+    Long authorId = boardService.getAuthorIdByBoardId(boardId);
+    return loginUser.getUserId().equals(authorId);
   }
 }
